+++ iModelCore/ECDb/ECDb/ECDbPch.h
@@ -37,6 +37,7 @@
 
 #include "ECDbSqlFunctions.h"
 
+#include "ECDbValidationChecks.h"
 #include "DbMappingManager.h"
 #include "DbMapValidator.h"
 #include "MapStrategy.h"