+++ iModelJsNodeAddon/api_package/ts/src/NativeLibrary.ts
@@ -434,6 +434,7 @@ export declare namespace IModelJsNative {
 
   interface SchemaImportOptions {
     readonly schemaLockHeld?: boolean;
+    readonly ecSchemaXmlContext?: ECSchemaXmlContext;
   }
 
   // ###TODO import from core-common
@@ -489,7 +490,7 @@ export declare namespace IModelJsNative {
     public executeTest(testName: string, params: string): string;
     public exportGraphics(exportProps: any/* ExportGraphicsProps */): DbResult;
     public exportPartGraphics(exportProps: any/* ExportPartGraphicsProps */): DbResult;
-    public exportSchema(schemaName: string, exportDirectory: string): SchemaWriteStatus;
+    public exportSchema(schemaName: string, exportDirectory: string, outFileName?: string): SchemaWriteStatus;
     public exportSchemas(exportDirectory: string): SchemaWriteStatus;
     public extractChangedInstanceIdsFromChangeSets(changeSetFileNames: string[]): ErrorStatusOrResult<IModelStatus, ChangedInstanceIdsProps>;
     public extractChangeSummary(changeCacheFile: ECDb, changesetFilePath: string): ErrorStatusOrResult<DbResult, string>;